Extract some related logic to separate module
`ApplicationController` is one of the biggest controllers. I guess it
makes sense to try keep it thin.
